+++ lib/sbmt/strangler/action_invoker.rb
@@ -7,30 +7,37 @@ require_relative "work_modes/replace"
 module Sbmt
   module Strangler
     class ActionInvoker
-      def initialize(action, rails_controller)
-        @action = action
+      attr_reader :rails_controller, :strangler_action, :feature_flags
+
+      def initialize(rails_controller:, strangler_action:, feature_flags:)
         @rails_controller = rails_controller
+        @strangler_action = strangler_action
+        @feature_flags = feature_flags
       end
 
-      delegate :track_params_usage, :track_work_mode, to: :@rails_controller
+      delegate :track_params_usage, :track_work_mode, to: :rails_controller
 
       def call
         track_params_usage
-        work_mode_class = choose_work_mode_class
-        track_work_mode(work_mode_class.name_for_metric)
-        work_mode_class.new(@action, @rails_controller).call
+        track_work_mode(work_mode_class.name.demodulize.underscore)
+        work_mode_class.new(
+          rails_controller:,
+          strangler_action:,
+          feature_flags:
+        ).call
       end
 
       private
 
-      def choose_work_mode_class
-        return Sbmt::Strangler::WorkModes::Replace if enabled?(@action.feature_flags.replace_work_mode)
-        return Sbmt::Strangler::WorkModes::Mirror if enabled?(@action.feature_flags.mirror_work_mode)
-        Sbmt::Strangler::WorkModes::Proxy
-      end
-
-      def enabled?(feature_name)
-        @rails_controller.flipper_feature_enabled_anyhow?(feature_name)
+      def work_mode_class
+        @work_mode_class ||=
+          if feature_flags.replace?
+            Sbmt::Strangler::WorkModes::Replace
+          elsif feature_flags.mirror?
+            Sbmt::Strangler::WorkModes::Mirror
+          else
+            Sbmt::Strangler::WorkModes::Proxy
+          end
       end
     end
   end