+++ lib/sbmt/strangler/work_modes/proxy.rb
@@ -7,13 +7,13 @@ module Sbmt
     module WorkModes
       class Proxy < Base
         def call
-          proxy_response = http_request(http_params)
-          render_proxy_response(proxy_response)
+          origin_response = http_request(http_params)
+          render_origin_response(origin_response)
         end
 
         private
 
-        delegate :http_params, :http_request, :render_proxy_response, to: :@rails_controller
+        delegate :http_params, :http_request, :render_origin_response, to: :rails_controller
       end
     end
   end